requirements.txt
pandas>=2.0,<2.3
pandas_ta==0.3.14b
numpy>=1.24,<2.0
yfinance>=0.2.40
notion-client>=2.2.1
PyYAML>=6.0.1
python-dotenv>=1.0.1